backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# Utils
beautifulsoup4==4.12.3

# AI
google-generativeai

# Optional
numpy
scikit-learn